requirements.txt
fastapi
uvicorn[standard]
google-cloud-speech
google-cloud-translate
google-cloud-dialogflow-cx
python-multipart 